--- carleton-pathways/src/components/searchBar/SearchBar.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function SearchBar() {
-  return (
-    <div>SearchBar</div>
-  )
-}
