+++ carleton-pathways/src/screens/homePage/HomePage.jsx
@@ -4,7 +4,7 @@ import CourseHeader from '../../components/courseHeader/CourseHeader'
 export default function HomePage() {
   return (
     <div class="flex justify-center m-4">
-    <CourseHeader></CourseHeader>
+    <CourseHeader/>
     </div>
   )
 }
